% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_security.R
\name{adjust_taxable_income_for_social_security}
\alias{adjust_taxable_income_for_social_security}
\title{Adjust taxable income for Social Security benefits}
\usage{
adjust_taxable_income_for_social_security(
  taxable_income,
  social_security_benefit
)
}
\arguments{
\item{taxable_income}{Taxable income}

\item{social_security_benefit}{Social Security benefit}
}
\value{

}
\description{
Reduces taxable income by half of social security benefits
}
\examples{
adjust_taxable_income_for_social_security(30000, 12000)
}
