% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_allocation.R
\name{allocate_assets}
\alias{allocate_assets}
\title{Allocate Assets}
\usage{
allocate_assets(
  total_amount = numeric(),
  annual_amount = numeric(),
  years_until_available = numeric()
)
}
\arguments{
\item{total_amount}{numeric}

\item{annual_amount}{numeric}

\item{years_until_available}{numeric}
}
\value{
list
}
\description{
Allocate Assets
}
\examples{
allocate_assets(100000, 20000, 10)
}
