% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_portfolio.R
\name{bridge_amount}
\alias{bridge_amount}
\title{Get bridge amount at age}
\usage{
bridge_amount(age, goal_annual_amount, return_rate = 0.04)
}
\arguments{
\item{age}{Current age}

\item{goal_annual_amount}{Goal amount to withdraw each year in early retirement}

\item{return_rate}{Return rate of the bridge account}
}
\value{

}
\description{
Get the amount needed to bridge your early retirement until age 65 based on
a withdrawal rate.
}
