% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_security.R
\name{calc_social_security_benefit}
\alias{calc_social_security_benefit}
\title{Calculate Social Security benefits}
\usage{
calc_social_security_benefit(lifetime_income, ss_bends)
}
\arguments{
\item{lifetime_income}{Lifetime income by year}

\item{ss_bends}{Data frame of social security bends: income thresholds and
benefit percents}
}
\value{

}
\description{
Calculate Social Security benefits
}
\examples{
calc_social_security_benefit(get_lifetime_income(25, 30000), ss_bends)
}
