% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_efficient_deferrals.R
\name{deferral_tax_bracket}
\alias{deferral_tax_bracket}
\title{Tax bracket in/over which to defer taxes}
\usage{
deferral_tax_bracket(future_income, tax_brackets, ss_benefit, years_savings)
}
\arguments{
\item{future_income}{A vector of future income until retirement}

\item{tax_brackets}{A table of tax brackets}

\item{ss_benefit}{A double of annual Social Security benefit in retirement}

\item{years_savings}{A double of years of savings needed to withdraw
in retirement}
}
\value{
A tax bracket from tax_brackets
}
\description{
Tax bracket in/over which to defer taxes
}
\examples{
deferral_tax_bracket(get_future_income(25, 30000), tax_brackets, 2000, 22)
}
