% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_CER.R
\name{extend_CER}
\alias{extend_CER}
\title{extend_CER}
\usage{
extend_CER(infl_vec, ...)
}
\value{
CER nuevo con proyección futura y una columna indicando qué valores son reales (previos) y qué valores proyectados
}
\description{
Devuelve el CER con proyección futura basada en una serie de inflaciones mensuales. Agrega tantos meses como elementos en el vector de inflación.
}
