% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_portfolio.R
\name{fastest_goal}
\alias{fastest_goal}
\title{Fastest goal total amount}
\usage{
fastest_goal(
  principle,
  maximum_contribution,
  age,
  goal_annual_amount,
  constant = 1.01,
  return_rate = 0.05
)
}
\arguments{
\item{principle}{The initial total amount}

\item{maximum_contribution}{The maximum annual contribution possible}

\item{age}{Your current age}

\item{goal_annual_amount}{The goal annual amount to withdraw in retirement}

\item{constant}{Optional constant to reduce short-term returns. Default determined from experimentation.}

\item{return_rate}{Optional long term rate of return. Default determined from experimentation.}
}
\value{
A number of years until growth and contributions reaches goal growth
}
\description{
How quickly can you reach the goal total amount, given a starting amount,
maximum annual contribution, current age, and goal annual amount.
}
\details{
Numerical approximation using Newton's method.
https://math.stackexchange.com/questions/925838/constructing-a-while-loop-in-r-for-newtons-method
}
