% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_minimum_distributions.R
\name{find_constant_withdrawal_rate}
\alias{find_constant_withdrawal_rate}
\title{Find Constant Withdrawal Rate}
\usage{
find_constant_withdrawal_rate(rmds, growth_rate = 1.03, decimals = 4)
}
\arguments{
\item{rmds}{A vector of required minimum distributions (0-100)}

\item{growth_rate}{A constant growth rate}

\item{increment}{The precision of the final withdrawal rate}
}
\value{

}
\description{
Finds a constant withdrawal rate that can be applied to a starting account
balance when required minimum distributions (RMDs) begin. This value enables
you to predict income from RMDs throughout retirement and keep the taxable
income below a goal tax threshold
}
\examples{
find_constant_withdrawal_rate(20:10)
}
