% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_minimum_distributions.R
\name{find_max_withdrawal}
\alias{find_max_withdrawal}
\title{Find Maximum Withdrawal}
\usage{
find_max_withdrawal(
  rmds,
  account_balance = 1,
  constant_withdrawal = 0,
  growth_rate = 1.03
)
}
\arguments{
\item{rmds}{A vector of required minimum distributions (between 0 and 100)}

\item{account_balance}{The starting account balance}

\item{constant_withdrawal}{A constant amount to withdraw each year}

\item{growth_rate}{An annual growth rate for the account balance}
}
\value{

}
\description{
Withdraws the maximum of required minimum distributions (RMDs) or a constant
withdrawal amount then grows the account by a constant growth rate.
}
\examples{
find_max_withdrawal(20:10)
}
