% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_minimum_distributions.R
\name{find_years_constant_withdrawals}
\alias{find_years_constant_withdrawals}
\title{Find the number of years possible with constant withdrawals}
\usage{
find_years_constant_withdrawals(
  withdrawal_amount,
  account_balance = 1,
  growth_rate = 1.03
)
}
\arguments{
\item{withdrawal_amount}{Annual withdrawal amount}

\item{account_balance}{Starting account balance}

\item{growth_rate}{Annual growth rate}
}
\value{

}
\description{
Finds the number of years of constant withdrawals from a growing account
before the account reaches 0.
}
\examples{
find_years_constant_withdrawals(0.04)
}
