% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_value.R
\name{future_value}
\alias{future_value}
\title{Calculate Future value}
\usage{
future_value(p, r, t)
}
\arguments{
\item{p}{numeric}

\item{r}{numeric}

\item{t}{numeric}
}
\value{
numeric
}
\description{
Calculate Future value
}
\examples{
future_value(2500,0.043,8/12)#2571.67
}
