% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_allocation.R
\name{get_amount_bonds}
\alias{get_amount_bonds}
\title{Get an amount of bonds to include in your portfolio}
\usage{
get_amount_bonds(
  total_amount = numeric(),
  annual_amount = numeric(),
  years_waiting = numeric(),
  constant = 1.01,
  return_rate = 0.05,
  ...
)
}
\arguments{
\item{total_amount}{Total amount in the portfolio. Numeric greater than 0.}

\item{annual_amount}{Annual amount to be withdrawn from the portfolio. Numeric greater than 0.}

\item{years_waiting}{Years until the annual amount will start being withdrawn. Numeric greater than or equal to 0.}

\item{constant}{Optional constant to reduce short-term returns. Default determined from experimentation.}

\item{return_rate}{Optional long term rate of return. Default determined from experimentation.}

\item{intercept}{Optional intercept for efficient bond allocation. Must be between 0 and 1. Defaults determined by experimentation.}

\item{slope}{Optional slope for efficient bond allocation. Must be between -1 and 0. Defaults determined by experimentation.}
}
\value{
An amount of bonds to include in your portfolio
}
\description{
Suggests an amount of bonds to include in your portfolio based on
the total value of your portfolio, the annual amount you'd like to withdraw,
and the number of years until withdrawals begin.
}
\examples{
get_amount_bonds(100000, 20000, 10)
}
