% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_portfolio.R
\name{get_goal_total_amount}
\alias{get_goal_total_amount}
\title{Get total amount needed to theoretically last forever}
\usage{
get_goal_total_amount(
  annual_amount,
  years_waiting,
  constant = 1.01,
  return_rate = 0.05
)
}
\arguments{
\item{annual_amount}{Annual amount to be withdrawn from the portfolio. Numeric greater than 0.}

\item{years_waiting}{Years until the annual amount will start being withdrawn. Numeric greater than or equal to 0.}

\item{constant}{Optional constant to reduce short-term returns. Default determined from experimentation.}

\item{return_rate}{Optional long term rate of return. Default determined from experimentation.}
}
\value{
Goal total amount
}
\description{
Given a goal annual amount and years waiting, what total amount of money
is needed to theoretically last forever?
}
