% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plainSrc.R
\name{get_index}
\alias{get_index}
\title{Get all members of an index}
\usage{
get_index(index_name, src = "yahoo", country = "germany")
}
\arguments{
\item{index_name}{The name of the Index}

\item{src}{The source that should be use (Yahoo or fcs)}

\item{country}{The country of the specified index (necessary if fcs is used as source)}
}
\description{
gets the the members of a given index such as ATX. The library does not
limit the number of indices as long as the used data sources are supporting
them
}
\details{
the method offers a source parameter to determine which
type of src should be used. The default one is yahoo.
Thereby, the members are taken from yahoo and the identifiers/sedol
are taken from stockmarketmba.com

The second type of implemented source is fcs. Here, only fcs is used
as this data source  uses symbol names that are incompatible with
stockmarketmba.com. So no identifier and sedol are provided so far for that
source. Be careful: fcs uses an open API key which might be overused!!!

Yahoo and fcs use completely different interfaces: for yahoo we were using
the data from the website while for fcs we were using the API

stockmarketmba.com doesn't provide identifiers for all shares and symbols
so it might be \code{NA}

Attention: fcs and yahoo use DIFFERENT symbols

AS DIFFERENT KIND OF SOURCES ARE USED, THE QUERIES ARE PRETTY SLOW
}
\examples{
get_index("DAX","fcs","germany")
get_index("ATX")
get_index("GDAXI") #identifiers are at least partly available
get_index("SOMA")
get_index("IXIC")
get_index("OMX")
}
