% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{get_lifetime_income}
\alias{get_lifetime_income}
\title{Get lifetime income}
\usage{
get_lifetime_income(
  current_age,
  current_income,
  starting_age = 18,
  retirement_age = 70
)
}
\arguments{
\item{current_age}{Current age}

\item{current_income}{Current annual income}

\item{starting_age}{Age when started working}

\item{retirement_age}{Age when stopped working}
}
\value{

}
\description{
Estimates lifetime income based on current age and income.
Used to estimate social security benefits.
}
\examples{
get_lifetime_income(25, 30000)
}
