% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_efficient_deferrals.R
\name{get_max_tax_bracket}
\alias{get_max_tax_bracket}
\title{Get maximum tax bracket}
\usage{
get_max_tax_bracket(future_income, tax_brackets)
}
\arguments{
\item{future_income}{A vector of future incomes}

\item{tax_brackets}{A table of tax brackets}
}
\value{

}
\description{
Finds the maximum tax bracket that will be paid with future incomes
}
\examples{
get_max_tax_bracket(c(10000, 50000), tax_brackets)
}
