% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_efficient_deferrals.R
\name{get_total_income_over}
\alias{get_total_income_over}
\title{Get total income over a threshold}
\usage{
get_total_income_over(income_threshold, future_income)
}
\arguments{
\item{income_threshold}{A double}

\item{future_income}{A vector of future incomes}
}
\value{
A double
}
\description{
Calculate total income over a threshold
}
\examples{
get_total_income_over(10000, c(15000, 20000))
}
