% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_portfolio.R
\name{get_years_until_depleted}
\alias{get_years_until_depleted}
\title{Get years until depletion}
\usage{
get_years_until_depleted(
  total_amount = numeric(),
  annual_amount = numeric(),
  years_waiting,
  constant = 1.01,
  return_rate = 0.05
)
}
\arguments{
\item{total_amount}{Total amount in the portfolio. Numeric greater than 0.}

\item{annual_amount}{Annual amount to be withdrawn from the portfolio. Numeric greater than 0.}

\item{years_waiting}{Years until the annual amount will start being withdrawn. Numeric greater than or equal to 0.}

\item{constant}{Optional constant to reduce short-term returns. Default determined from experimentation.}

\item{return_rate}{Optional long term rate of return. Default determined from experimentation.}
}
\value{
positive numeric or Inf
}
\description{
Calculate how many years a portfolio will last until it is fully depleted
}
\examples{
get_years_until_depleted(10000, 1000, 5)
}
