% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLecaps.R
\name{getLecaps}
\alias{getLecaps}
\title{getLecaps}
\usage{
getLecaps(...)
}
\value{
Vector con los tickers de las Lecaps vigentes y resto de info.
}
\description{
Devuelve las Lecaps
}
\examples{
getLecaps()
}
