% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_portfolio.R
\name{growth_and_contributions}
\alias{growth_and_contributions}
\title{Growth with contributions}
\usage{
growth_and_contributions(principle, contribution, years, return_rate = 0.05)
}
\arguments{
\item{principle}{The initial amount}

\item{contribution}{The annual contribution}

\item{years}{The number of years of growth and contribution}

\item{return_rate}{The rate of return each year}
}
\value{
The final amount
}
\description{
Growth of an account with both growth and contributions over time.
http://www.moneychimp.com/articles/finworks/fmbasinv.htm
}
