% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income_trajectories}
\alias{income_trajectories}
\title{Income trajectory data}
\format{
An object of class 'lm'
}
\source{
\href{https://dqydj.com/income-percentile-by-age-calculator/}{DQYDJ}
}
\usage{
data(income_trajectory)
}
\description{
Data of income by age and income percentile
}
\examples{
data(income_trajectories)
lm(income ~ age + percent, income_trajectories)
}
\keyword{datasets}
