% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflacionbe.R
\name{inflacionbe}
\alias{inflacionbe}
\title{Calcula la inflación BE}
\usage{
inflacionbe(fecha, ...)
}
\arguments{
\item{fecha}{Fecha de cálculo}

\item{...}{Argumentos adicionales como server y port}
}
\value{
Dataframe con la inflación BE
}
\description{
Calcula la inflación BE a una fecha puntual. Para ejecutar un rago, deberá ser llamada cada vez y concatenar.
}
\examples{
\dontrun{
inflacionbe("2021-01-01")
inflacionbe("2021-01-01", server = "medina", port = 14493)
}
}
