% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_portfolio.R
\name{lowest_contribution}
\alias{lowest_contribution}
\title{Lowest contribution with growth and contribution}
\usage{
lowest_contribution(principle, years, final_amount, return_rate = 0.05)
}
\arguments{
\item{principle}{The initial amount}

\item{years}{The number of years of growth and contribution}

\item{final_amount}{The final amount after years of growth and contribution}

\item{return_rate}{The rate of return each year}
}
\value{
A contribution rate
}
\description{
What is the lowest contribution rate that will reach a given final amount
given a starting and final amount, years, and return rate?
}
