% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planning.R
\name{plan_retirement_allocation}
\alias{plan_retirement_allocation}
\title{Plan retirement allocation}
\usage{
plan_retirement_allocation(
  birth_year,
  age,
  starting_amount,
  annual_withdrawal,
  limit_change,
  contribution_or_withdrawal_amount,
  bond_rates,
  stock_rates
)
}
\arguments{
\item{birth_year}{Year of birth}

\item{age}{Current age}

\item{starting_amount}{Current amount of money saved for retirement}

\item{annual_withdrawal}{Annual amount desired to be withdrawn in retirement}

\item{limit_change}{Should changes in the asset allocation be limited to contributions and withdrawals?}

\item{contribution_or_withdrawal_amount}{A vector of amount to contribute or withdraw}

\item{bond_rates}{A vector of changes in bond values over time}

\item{stock_rates}{A vector of changes in stock values over time}
}
\value{
A graph of the theoretical performance of the bond allocation model
}
\description{
Predict and plan suggested retirement allocation using the bond allocation model.
}
\examples{
plan_retirement_allocation(1991, 21, 1000, 40000, TRUE, rep(6000, 10), rep(1.02, 10), rep(1.05, 10))
}
