% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rmds}
\alias{rmds}
\title{Required minimum distributions (RMDs)}
\format{
A double vector
}
\source{
\href{https://www.irs.gov/publications/p590b#en_US_2019_publink1000231258}{RMDs}
}
\usage{
data(rmds)
}
\description{
Required minimum distributions for deferred taxable income in retirement in
2021
}
\examples{
data(rmds)
}
\keyword{datasets}
