% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smooth_income_by_percent_age}
\alias{smooth_income_by_percent_age}
\title{Smoothed income trajectory data}
\format{
An object of class 'scam', 'glm', 'lm'
}
\source{
\href{https://dqydj.com/income-percentile-by-age-calculator/}{DQYDJ}
}
\usage{
data(smooth_income_by_percent_age)
}
\description{
Smoothed shape-constrained additive model (SCAM) of
income by age and income percentile
}
\examples{
data(smooth_income_by_percent_age)
predict(
  smooth_income_by_percent_age,
  data.frame(percent = 0.5, age = 25)
)
}
\keyword{datasets}
