% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smooth_percent_by_income_age}
\alias{smooth_percent_by_income_age}
\title{Smoothed income percentile data}
\format{
An object of class 'scam', 'glm', 'lm'
}
\source{
\href{https://dqydj.com/income-percentile-by-age-calculator/}{DQYDJ}
}
\usage{
data(smooth_percent_by_income_age)
}
\description{
Smoothed shape-constrained additive model (SCAM) of
income percentiles by income and age
}
\examples{
data(smooth_percent_by_income_age)
predict(
  smooth_percent_by_income_age,
  data.frame(income = 30000, age = 25)
)
}
\keyword{datasets}
