% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_security.R
\name{social_security_bends}
\alias{social_security_bends}
\title{Social Security Bends}
\usage{
social_security_bends(monthly_income, ss_bends)
}
\arguments{
\item{monthly_income}{Monthly income from top 35 average annual incomes}

\item{ss_bends}{Data frame of social security bends: income thresholds and
benefit percents}
}
\value{

}
\description{
Calculate social security (SS) benefits from average monthly income and
SS bends
}
\examples{
social_security_bends(2000, ss_bends)
}
