% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_security.R
\name{social_security_benefit}
\alias{social_security_benefit}
\title{Estimate Social Security benefit}
\usage{
social_security_benefit(current_age, current_income, ss_bends)
}
\arguments{
\item{current_age}{Current age}

\item{current_income}{Current annual income}

\item{ss_bends}{Data frame of social security bends: income thresholds and
benefit percents}
}
\value{

}
\description{
Estimate Social Security benefit
}
\examples{
social_security_benefit(25, 30000, ss_bends)
}
