% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss_bends}
\alias{ss_bends}
\title{Social security bends}
\format{
An object of class 'data.frame'
}
\source{
\href{https://www.ssa.gov/oact/COLA/piaformula.html}{PIA formula}
}
\usage{
data(ss_bends)
}
\description{
Dataframe of social security bends and percent benefits for 2021
}
\examples{
data(ss_bends)
}
\keyword{datasets}
