% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasasLecap.R
\name{tasasLecap}
\alias{tasasLecap}
\title{Tasas Lecap}
\usage{
tasasLecap(df, settle = "t+1", ...)
}
\arguments{
\item{df}{data.frame con los datos de los bonos}

\item{settle}{fecha de liquidación de la operación. Puede ser "t+0" o "t+1" versus la fecha del precio}

\item{...}{argumentos adicionales para dbGetTable}
}
\value{
data.frame recibido con el agregado de todos los calculos realizados
}
\description{
Calcula las tasas y duration de las lecap. Obtiene de la tabla "lecaps" los
vencimientos y valores finales de cada lecap de la base de datos. Calcula las tasas directas,
tna, tea, tem, tna360, tea360, tem360 y duration de cada lecap.
Necesita que el df tenga date, price, ticker
}
\examples{
\dontrun{
tasasLecap(df, settle = "t+1")
}

}
