% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasasTamar.R
\name{tasasTamar}
\alias{tasasTamar}
\title{Tasas Tamar}
\usage{
tasasTamar(df, settle = "t+1", ...)
}
\arguments{
\item{df}{data.frame con los datos de los bonos}

\item{settle}{fecha de liquidación de la operación. Puede ser "t+0" o "t+1" versus la fecha del precio}

\item{...}{argumentos adicionales para dbGetTable}
}
\value{
data.frame recibido con el agregado de todos los calculos realizados
}
\description{
Calcula las tasas y duration de las tamar. Obtiene de la tabla "tamar" los
vencimientos de cada titulo a tamar de la base de datos. Calcula los vpv, tasas directas,
tna, tea, tem, tna360, tea360, tem360 y duration de cada lecap.
Necesita que el df tenga date, price, ticker
}
\examples{
tasasLecap(df, settle = "t+1")


}
