% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax_brackets}
\alias{tax_brackets}
\title{Tax brackets}
\format{
An object of class 'data.frame'
}
\source{
\href{https://www.irs.gov/newsroom/irs-provides-tax-inflation-adjustments-for-tax-year-2020}{IRS 2020}
}
\usage{
data(tax_brackets)
}
\description{
Dataframe of tax brackets and tax rates for 2020
}
\examples{
data(tax_brackets)
}
\keyword{datasets}
