% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_efficient_deferrals.R
\name{tax_efficient_deferrals}
\alias{tax_efficient_deferrals}
\title{Tax efficient deferrals}
\usage{
tax_efficient_deferrals(
  current_age,
  future_income,
  income_trajectories = income_trajectories_2021,
  tax_brackets = tax_brackets_2021
)
}
\arguments{
\item{income_trajectories}{Table of income trajectories}

\item{tax_brackets}{Table of tax brackets}

\item{income}{Current income}

\item{age}{Current age}
}
\value{

}
\description{
Tax efficient deferrals
}
