% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_efficient_deferrals.R
\name{total_deferred_savings}
\alias{total_deferred_savings}
\title{Total deferred savings}
\usage{
total_deferred_savings(
  current_age,
  current_income,
  ss_bends,
  rmds,
  tax_brackets,
  growth_rate = 1.03
)
}
\arguments{
\item{current_age}{Current age}

\item{current_income}{Current income}

\item{ss_bends}{A table of social security bends and benefit rates}

\item{rmds}{A vector of required minimum distributions}

\item{tax_brackets}{A table of tax brackets and tax rates}

\item{growth_rate}{Annual growth rate}
}
\value{

}
\description{
Calculates the total savings needed to maximize tax deferred withdrawals
}
\examples{
total_deferred_savings(25, 30000, ss_bends, rmds, tax_brackets)
}
