% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_minimum_distributions.R
\name{total_years_deferred_savings}
\alias{total_years_deferred_savings}
\title{Total years of deferred savings}
\usage{
total_years_deferred_savings(rmds, growth_rate = 1.03)
}
\arguments{
\item{rmds}{A vector of required minimum distributions (0-100)}

\item{growth_rate}{Annual growth rate}
}
\value{

}
\description{
Calculates the number of years of savings you would need in your
tax-deferred account to withdraw the same amount each year, assuming the
account also continues to grow. This value can be multiplied by the amount
withdrawn each year to determine an initial total amount of savings.
}
\examples{
total_years_deferred_savings(rmds)
}
