% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evol_fechas.R
\name{evol_fechas}
\alias{evol_fechas}
\title{Evolución de un campo entre dos fechas}
\usage{
evol_fechas(df, fecha1, fecha2, campo, campo_fecha = "date")
}
\arguments{
\item{fecha1}{Fecha inicial para la comparación.}

\item{fecha2}{Fecha final para la comparación.}

\item{campo}{Nombre del campo a comparar.}

\item{campo_fecha}{Nombre del campo que contiene las fechas (por defecto "date").}
}
\value{
Dataframe con la evolución del campo entre las dos fechas, incluyendo la variación porcentual.
}
\description{
Esta función permite comparar un campo específico entre dos fechas en un dataframe.
#' @param df Dataframe que contiene los datos.
}
\examples{
evol_fechas(df, fecha1= "2025-07-15", fecha2 = "2025-07-31", campo = "price", campo_fecha = "date")

}
