% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bandas.R
\name{get_bandas}
\alias{get_bandas}
\title{get_bandas}
\usage{
get_bandas(end_date = "2027-12-31")
}
\value{
Un data frame con las bandas superior e inferior y la fecha.
}
\description{
Devuelve las bandas superior e inferior según la "FASE 3" del programa monetario del gobierno Milei. Comienza el 14-abr-25 con 1000 y 1400. Ajusta al 1% mensual
}
