% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_efficient_deferrals.R
\name{get_future_income}
\alias{get_future_income}
\title{Get future income trajectory}
\usage{
get_future_income(current_age, current_income, retirement_age = 70)
}
\arguments{
\item{current_age}{Current age}

\item{current_income}{Current income}
}
\value{
Vector of future incomes
}
\description{
Assumes income will increase from until 40 then plateau
}
