% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{get_future_income}
\alias{get_future_income}
\title{Get future income trajectory}
\usage{
get_future_income(current_age, current_income, retirement_age = 70)
}
\arguments{
\item{current_age}{Current age}

\item{current_income}{Current income}

\item{retirement_age}{Age when stopped working}
}
\value{
Vector of future incomes
}
\description{
Estimates future income until retirement from current age and income.
Used to estimate total lifetime income to determine optimal tax brackets
for tax-deferral.
}
\examples{
get_future_income(25, 30000)
}
