% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{get_income_percentile}
\alias{get_income_percentile}
\title{Income percentile}
\usage{
get_income_percentile(current_age, current_income)
}
\arguments{
\item{current_age}{Current age}

\item{current_income}{Current annual income}
}
\value{
Estimated income percentile
}
\description{
Estimates income percentile given current age and income.
Used to estimate future earnings.
}
\examples{
get_income_percentile(25, 30000)
}
