% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_efficient_deferrals.R
\name{get_income_percentile}
\alias{get_income_percentile}
\title{Income percentile}
\usage{
get_income_percentile(age, income)
}
\value{
Table of income percentiles
}
\description{
Income percentile
}
