% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLecaps.R
\name{getLecaps}
\alias{getLecaps}
\title{getLecaps}
\usage{
getLecaps(date = Sys.Date(), ...)
}
\arguments{
\item{date}{Fecha en formato "YYYY-MM-DD". Por defecto la fecha actual.}

\item{...}{Parametros de conexion a la base de datos}
}
\value{
Vector con los tickers de las Lecaps vigentes
}
\description{
Devuelve las Lecaps vigentes segun fecha provista
}
\examples{
getLecaps("2024-08-10")
getLecaps()
}
