% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflacionbe.R
\name{inflacionbe}
\alias{inflacionbe}
\title{Calcula la inflaci칩n BE}
\usage{
inflacionbe(fecha, ...)
}
\arguments{
\item{fecha}{Fecha de c치lculo}

\item{...}{Argumentos adicionales como server y port}
}
\value{
Dataframe con la inflaci칩n BE
}
\description{
Calcula la inflaci칩n BE
}
\examples{
inflacionbe("2021-01-01")
inflacionbe("2021-01-01", server = "medina", port = 14493)
}
