% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_efficient_deferrals.R
\name{tax_efficient_deferrals}
\alias{tax_efficient_deferrals}
\title{Tax efficient deferrals}
\usage{
tax_efficient_deferrals(
  current_age,
  tax_brackets,
  current_savings_tax_deferred,
  future_income
)
}
\arguments{
\item{current_age}{Current age}

\item{tax_brackets}{Table of tax brackets}

\item{current_savings_tax_deferred}{Current amount of tax-deferred savings}

\item{future_income}{Vector of future income until retirement}
}
\value{

}
\description{
Tax efficient deferrals
}
