% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_efficient_deferrals.R
\name{tax_efficient_deferrals}
\alias{tax_efficient_deferrals}
\title{Tax efficient deferrals}
\usage{
tax_efficient_deferrals(
  current_age,
  current_income,
  ss_bends,
  rmds,
  tax_brackets
)
}
\arguments{
\item{current_age}{Current age}

\item{current_income}{Current annual income}

\item{ss_bends}{A table of social security bends and benefit rates}

\item{rmds}{A vector of required minimum distributions}

\item{tax_brackets}{A table of tax brackets and tax rates}
}
\value{

}
\description{
Estimate the tax bracket in/above which you should defer taxes
}
\examples{
tax_efficient_deferrals(25, 30000, ss_bends, rmds, tax_brackets)
}
